test framework setup
