+++ include/program.hpp
@@ -1,0 +1,8 @@
+#ifndef PROGRAM_H
+#define PROGRAM_H
+
+class Program {
+    public:
+        static int program();
+};
+#endif
