+++ src/test.cpp
@@ -1,0 +1,39 @@
+//  (C) Copyright Gennadiy Rozental 2005.
+//  Distributed under the Boost Software License, Version 1.0.
+//  (See accompanying file LICENSE_1_0.txt or copy at 
+//  http://www.boost.org/LICENSE_1_0.txt)
+
+//  See http://www.boost.org/libs/test for the library home page.
+
+// Boost.Test
+
+// each test module could contain no more then one 'main' file with init function defined
+// alternatively you could define init function yourself
+#define BOOST_TEST_MAIN
+#include <boost/test/included/unit_test.hpp>
+
+//____________________________________________________________________________//
+
+// most frequently you implement test cases as a free functions with automatic registration
+BOOST_AUTO_TEST_CASE( test1 )
+{
+    // reports 'error in "test1": test 2 == 1 failed'
+    BOOST_CHECK( 2 == 1 );
+}
+
+//____________________________________________________________________________//
+
+// each test file may contain any number of test cases; each test case has to have unique name
+BOOST_AUTO_TEST_CASE( test2 )
+{
+    int i = 0;
+
+    // reports 'error in "test2": check i == 2 failed [0 != 2]'
+    BOOST_CHECK_EQUAL( i, 2 );
+
+    BOOST_CHECK_EQUAL( i, 0 );
+}
+
+//____________________________________________________________________________//
+
+// EOF
